Refactor App_Brake.h
