+++ boards/FSM/Inc/Io/Io_Brake.h
@@ -3,7 +3,7 @@
 #include <stdbool.h>
 
 /**
-* Check if the brake is actuated
-* @return True if the brake is actuated, false if the brake is not actuated
+ * Check if the brake is actuated
+ * @return True if the brake is actuated, false if the brake is not actuated
  */
 bool Io_Brake_IsActuated(void);
